+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPQH7SQS7F56BDPS9DKW63G4"
+        "01KPQP8K7HRKSCCP3GJEWVK8DJ"
     )
     r"""The unique identifier of the data source"""