+++ packages/orq-rc/RELEASES.md
@@ -7129,3 +7129,13 @@ Based on:
 - [python v4.8.0-rc.35] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.35] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.35 - packages/orq-rc
+
+## 2026-04-21 09:37:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.36] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.36] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.36 - packages/orq-rc